requirements.txt
autopep8==1.5.2
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
pycodestyle==2.5.0
python-dateutil==2.8.1
requests==2.23.0
six==1.14.0
timeloop==1.0.2
urllib3==1.25.9
